fix(capture): read NL80211_BSS_SIGNAL_MBM as signed, and lint every platform
golangci-lint on macOS never compiles capture_linux.go or capture_windows.go,
so three gosec findings reached CI that no local run could have caught. make
lint now runs the pure-Go backends under GOOS=linux and GOOS=windows too.

The real one of the three: the signal attribute was decoded with Uint32 and
converted, where nl80211 defines it as signed. Every reading is negative, so
the unsigned form would have turned each one into a large positive number the
moment the conversion stopped being reinterpreted as two's complement.
Decoder.Int32 says what is meant and removes the conversion. Re-verified
against real hardware: 11 networks, values matching iw exactly.

The other two were test fixtures narrowing a length; both bounds are checked
now, which is what gosec wants and what the element format actually allows.

+++ internal/capture/capture_linux.go
@@ -372,7 +372,10 @@ func networkFromBSS(data []byte, seen time.Time) (wifi.ScannedNetwork, bool, err
 				case nl80211BSSFrequency:
 					freqMHz = int(nad.Uint32())
 				case nl80211BSSSignalMBM:
-					signalDBm = int(int32(nad.Uint32())) / mBmPerDBm
+					// Int32, not Uint32: NL80211_BSS_SIGNAL_MBM is signed, and
+					// reinterpreting the unsigned form would turn every real
+					// (negative) reading into a huge positive one.
+					signalDBm = int(nad.Int32()) / mBmPerDBm
 				case nl80211BSSCapability:
 					capability = nad.Uint16()
 				case nl80211BSSInformationElements:
